Add general test if a string results in value
For any type X that offers std::ostream << x, for X x, we have a struct
of a std::string and a X to be used in boost data tests.
There is a general, templated struct string_to_x, to use.

Overloads the necessary methods such that boost can handle this.
Works only for value types.